requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
python-docx==1.1.0
requests==2.31.0
oauthlib==3.2.2
python-dotenv==1.0.0